+++ packages/components/src/components/inline-notification/inline-notification.tsx
@@ -94,7 +94,7 @@ export class InlineNotification {
             hideLabel={true}
             compact={true}
             onClick={this.dismissNotification}
-            aria={{ 'aria-description': headingText }}
+            {...(headingText ? { aria: { 'aria-description': headingText } } : {})}
           >
             Close notification
           </PrefixedTagNames.pButton>